p3_app/requirements.txt
alembic==1.6.5
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
xlrd==2.0.1
